INSTALL.md: Split apt-get line for Node.js
Apparently, depending on version and/or configuration, "apt-get
install" will sometimes not install anything when it is presented with
a mixture of existing and non-existing package names. To make sure the
instructions in INSTALL.md work in all cases, give a separate apt-get
invocation for the nodejs-legacy package instead of one for all the
Node.js packages.