INSTALL.md: Move JSON::XS to the database section
The JSON::XS Perl module is required by a PL/Perl database function;
it must therefore be available to the Postgres server, unlike the
other (local::lib) modules. INSTALL.md used to discuss this in the
Perl dependencies section, though this recently led to a
misunderstanding when a user was setting up a separate database
server. Move this point to the database section instead.