Move entity serialization from JSONSerializer to TO_JSON methods
This fixes bugs introduced in 02cda98. Because the leading underscores got
removed, JSON::2::WS would pick up and use those JSONSerializer methods.

That could be fixed by adding back the underscores, but there's already been
parallel, ongoing work in the react-server branch to move entity serialization
to TO_JSON methods. This is convenient because it's common to put entities
in the stash for templates to use, and encoding the stash as JSON just becomes
a matter of enabling convert_blessed on the encoder (see
search.cpan.org/perldoc?JSON).

And by moving the addition of certain JSON keys to specific Entity::Role
packages, this also brings some added consistency to our JSON output, which is
another long-term goal. Existing keys were mostly untouched, though, to keep
compatibility throughout. An exception is that the parent_* attributes on areas
were replaced with a single 'containment' list.