Move knockout to package.json
The problem with this previously was that we're using a git version of
knockout. We couldn't upgrade to 3.2 because it introduced some bugs, and
couldn't downgrade to 3.1 because we're using the "components" feature.

I identified the exact commit we're using, but it didn't work to use directly
because knockout's package.json is broken in that commit.

I created a separate branch at mwiencek/knockout to host the exact version we
use.