Skip bad empty AC field check
The purpose of this code was to have some generic check for all required
fields on the page. If any were empty, the submit button was to be disabled.

This didn't work in practice for AC fields, because the main autocomplete
input for those (where the required attribute is set) doesn't tell whether an
entity is actually selected. Hidden input fields are added separately by
knockout. Which also means we can't easily attach an event to them that hooks
into the code here.

Skip artist credits here.