Fix bad eval of ShowDBDefs output in compile_resources.pl
Double quotes are needed around "$()" to prevent word splitting. This fixes
compile_resources.pl during chef provisioning.

The escaping in compile_resources.pl was getting kind of silly given that
it's a shell script embedded inside a string inside a perl script, so I just
turned it into a shell script.

compile_resources.pl will stick around until it's convenient to remove. The
use of `system` was replaced with `exec` to capture the exit code of the
shell script.