Replace dynamic require() for jed data with static import
There're some problems with the current implementation: it depends on a global
variable (window.MB_LANGUAGE), and then does a dynamic require() based on that
global.

We'll probably want to use ES6 import statements now that they're finalized,
which are not expressions and can't do dynamic imports. While that may seem
limiting, they can be statically analyzed that way, which is good.

One downside is that for simplicity, this requires creating a jed-data file for
English that exports an empty, plain object, but I don't see that as a huge
deal.