MBS-8514: relationship credits can't be edited
This issue was introduced in 88c0e9358d8c8cf81d5298dd0494ae58c7833622. While
that commit's reasoning is correct, Edit::Relationship::Edit was using
`exists` to mean "including entity credits," because otherwise, the
relationship being edited would always be a duplicate of itself if only the
credits were changed.

This modifies `exists` to return the ID of the relationship that exists, and
makes Edit::Relationship::Edit check if the ID differs from the relationship
being edited.