Fix MBS-8514 again
We need to add the same behavior described in cb510c5 to `accept`, not just
`initialize`. Annoyingly, the test I added in cb510c5 would have caught this
if it wasn't for a randomly-incorrect attribute_count column in
t/sql/edit_relationship_edit.sql, which somehow allowed the test to pass.