Remove MEMCACHED_NAMESPACE from redis args
This is just confusing and unnecessary. In production we don't use any
memcached namespace (because we override the entire CACHE_MANAGER_OPTIONS
block), but MEMCACHED_NAMESPACE retains its default value there, so it is
apparently used for redis. Therefore we can't set MEMCACHED_NAMESPACE to an
empty string in chef without also affecting redis.