Remove wget test from fabfile
Now that we've set a spawn-interval on the start_server process, this doesn't
really test what it's intended to. We'd have to increase the sleep() amount
even further to test this properly, which would in turn slow down deployment
significantly.

This also isn't really useful in general, because if the homepage didn't load,
we'd immediately get tons of ISE emails.