Enable rsync full export to taz server
It was set up for scooby server (RIP), so make changes to re-enable full export via rsync on taz (new ftp).

Note: syntax changed a bit because taz is using rrsync helper on its side, the `dummy` target is not needed anymore.
On taz, a user exists for this purpose called `ftp-data`.