Catch generic request timeouts like database timeouts
When a request exceeded the allowable time, the server process was
immediately killed. The user was shown a plain nginx error page.

Instead of this, throw an exception when a timeout occurs, just like
when a database query timed out. The exception can be caught and a
more meaningful error page be displayed. This also makes the
return-to-search-form feature of the edit search less fragile (it only
worked when the query timed out before the request did, MBS-8382).