Add a proper Catalyst view class for the Node template renderer
Fixes e.g. “Couldn't render template "file error - artist/show.tt: not found"”
errors that occur whenever an action detaches to /error_404; that action
doesn't have a `NewTemplate` attribute, but it wouldn't matter even if it did,
because only an attribute on the original action can affect the result at that
point, given the way the current code is written.

The `NewTemplate` attribute setup is flawed, so a `View::Node` class is added
instead, which an action can easily switch to by setting
`$c->stash->{current_view}`.