Make it possible to suppress certain error emails
Catalyst::Plugin::ErrorCatcher does not offer any option to turn off
error reporting for a particular error. It can only be enabled or
disabled globally at server startup time.

Add an ErrorCatcherEmailWrapper module that can be used as an emitter
class for ErrorCatcher and is a wrapper around its Email emitter.
Depending on a package variable that can be changed temporarily via
local, an error is forwarded to C::Pl::ErrorCatcher::Email or not.