Simplify debugging output for when inserting an edit fails
When inserting an edit failed with an exception, a new exception was
thrown containing the body (POST) parameters of the request and a dump
of the original exception. However, the request parameters are shown
anyway both on the ISE page and in the error email, and the dump of a
complex exception object is far too verbose.

Don't include the request parameters in the new exception, and just
show the stringified original exception together with its class name.