Moose::Util::does_role seems to work better than ->does in 5.18 (?)
Should fix MBS-8145 on pino.

Thanks to @uklauer for suggesting this. I'm not able to narrow down why this
works, but pino and I both have perl 5.18, where this fixes things, Ulrich has
5.20, where the existing code worked for him, and we both have the same version
of Moose. There are a lot of other factors not accounted for, but that is the
most obvious one to me.