Fix autocomplete for entities without a gid (MBS-7873, MBS-8051)
For MBS-7873, we now dedupe recent entities by id if the entity type doesn't
have mbids. (entities.json is imported to determine this.)

For MBS-8051, the issues arose because `clearSelection` only cleared the
entity if one was "selected," and it determined that based on whether the
current entity data had a gid. Now it looks at the id instead.