MBS-8681: Let edits again expire after 7, not 0 days
The changes in commit 909831ea267c19d7355db8a36a0d6f7ceba81e89 changed
the default number of days an edit was to stay open to zero, because
the "days" method on a Duration doesn't count days that are already
included for "weeks". Use "in_units" instead, which allows control
over which units are to be considered.