Use the link_artist macro in artist credit diffs
The Diff plugin had its own method for creating a link to an artist
that was not equivalent to the link_artist template macro. In
particular, name variations were not marked as such (MBS-7844).

Remove the separate implementation and call the macro instead. Extend
the macro to support already-encoded HTML input, like link_recording.