MBS-8698: Be more flexible in identifying JSON-LD requests
A request for the JSON-LD for an entity was only honoured if its HTTP
Accept header consisted of the exact string "application/ld+json".
Even adding a q value or a second acceptable MIME type caused the
server to fall back to the standard HTML page.

This commit makes the server look for any occurrence of the JSON-LD
type in the Accept header, which will fix the common scenarios
mentioned above. However, it might cause new problems if clients use
valid but unusual Accept headers (e.g. containing both "text/html" and
"application/ld+json", with the former being preferred).