Don't exit if the overall script is running
The main reason we were doing this was so that
`overall_sitemaps_replication_sequence` didn't get updated in the middle of
an incremental script run, since that would affect the data it outputs.

That doesn't really seem like a problem, though. The incremental script still
has to process all the packets anyway, and the sitemaps (and sitemap index)
have timestamps saying when things were updated, so it's always clear which
files are more recent.

Exiting to let the overall script finish is another reason the incremental
script was always behind on processing the packets, too.

One remaining issue was that the sitemap index file was only read once and then
cached, but now it may now be updated in the meantime. We fix that here and use
flock to protect against concurrent reads/writes that could cause data
inconsistency.