Escape </ within <script> tags
It's illegal for </ followed by a name to appear within a script tag:
http://www.w3.org/TR/REC-html32-19970114#script

Currently, it's possible to tag something as </script> and close the script
block on the page where it's injected.

This commit adds a closing_tag_escape macro to replace all occurrences of </
with <\/ in a value. (Encoding HTML entities within CDATA would be incorrect
and break things.) Right now, it's only used for JSON, which is fine because
the sequence </ is only legal inside quoted keys/values.

I believe the JSON.Escape plugin for TT also handles this. However, all of the
occurrences being changed here are already encoded as JSON strings, whereas
the .json vmethod expects a hash or array ref. And even if we changed the code
to pass refs to the templates, we'd have to make sure all of the same JSON
flags are used (many depend on allow_blessed, convert_blessed, utf8, etc.).