Avoid double-encoding in <title> tags
Commit 7c99b190fd6148db5d203a8cf39c4d235cf0cd96, which made the main
layout template encode the contents of the <title> tag, exposed some
places in individual templates where this was inconsistently already
done, now causing double-encoding. (One of these was MBS-8702.)

Remove the encoding from the individual templates. In one case where
the same string was used for title and heading, encode the heading.