Split both before and after separator for diffs
The existing implementation of _split_text for use by diff_side and
diff_html_side worked by inserting a special marker in front of each
occurrence of the separator, then splitting the string at these
markers. This meant that, e.g., "bar baz" and "foo bar baz" with a
whitespace separator got split into ["bar", " baz"] and ["foo",
" bar", " baz"], respectively; the diff algorithm then found "bar" and
" bar" to be unrelated and marked them as a change (MBS-5205).

An earlier attempt to fix this changed all whitespace separators into
zero-width word boundaries (\b), so that the input from above would
have been split into ["bar", " ", "baz"] and ["foo", " ", "bar", " ",
"baz"], respectively, giving a more plausible diff. However, that
attempt (commit 89afb8383ca8daa61b0541eb461203e31cbcf013) was reverted
the same day it was merged, for unknown reasons.

The new approach here accomplishes the same result using a special
feature of Perl's split: Any text matched by capture groups in the
separator regex will be put into the result list, as a separate item.
This is not only simpler than the existing implementation in that it
doesn't need to resort to a cryptographic hash function to find a
collision-free marker; unlike the previous improvement attempt, it
works with all kind of separators instead of only word boundaries, and
the required changes are localized to a single function. As a minor
disadvantage (shared with the previous attempt), a diff of two
unrelated multi-word strings will leave corresponding spaces between
words unmarked; this will be addressed in a later commit.