Prepare for moving sitemaps to lolo
Sitemaps will be built on a dedicated slave DB on lolo, instead of on astro
using the production DB.

We can't as easily dump the sitemaps schema anymore with this setup. It's
questionable how useful those dumps are right now, given how behind they are,
and you can still consume the changed timestamps from the sitemaps themselves.