Get rid of the display_url macro
Almost everywhere the display_url macro was used, its parameter was a
URL entity; it was therefore equivalent to calling url.pretty_name. In
the only place where it was used on a plain URI object (the XML link
on entity details pages), it had no particular effect either because
that URI never contains any percent-encoded characters. It can
therefore safely be removed.