MBS-8683: Decode internationalized domain names for display
Percent-encoded UTF-8 in the local part of URLs has long been decoded
for display, but punycoded internationalized domain names were shown
as raw ASCII.

Use the existing IRI support of the URI module to decode all parts of
URLs at the same time; add some new attributes (so that the decoding
only has to be done once) to the URL base class for access to the
decoded URL or just the decoded local part and for checking whether
the local part was in some non-UTF-8 legacy encoding.

Finally, make pretty_name use the decoded form, as well as two
templates that don't use a (possibly overriden) pretty_name.