Revert URL::Show test to again expect a decoded URL
The previous commit made the heading on the URL entity page again show
the form of the URL with UTF-8 characters decoded instead of the
pure ASCII form. Change the corresponding test back to expecting the
decoded form.

This mostly reverts commit e395099e4fd291370a88724f880ee4e149b65496,
but keeps the new test that was added there.