Check early whether there really is a logged-in user
The user_exists method does not actually try to restore the user; if
there is a valid session, it will return true, even if the owner of
the session has been deleted in the meantime. Check early for this
condition (by calling the user method instead of user_exists). This
results in a clearer error message than when the issue is detected
later inside the Redis access try/catch block.