MBS-8688: Use consistent order for entity types on tag pages
For user tag pages, the entity types were sorted using the <=>
operator. Since this operator treats its operands as numbers, the
strings "release", "recording" etc. were all the same to it, having
the numerical value 0. As Perl's sort is not guaranteed to be stable,
the resulting order was unpredictable.

Use cmp instead, the operator for lexicographic comparison of strings.