MBS-7241: Make cache transactional
This ensures that an entity being deleted can't be re-added to the cache
in a concurrent transaction before ours commits.

`pg_advisory_xact_lock` is used before cache deletion, and
`pg_try_advisory_xact_lock` (which is non-blocking) before adding things to the
cache. If the second case fails to obtain the lock, we don't add anything to
the cache. Both locks are automatically released at the end of the transaction.

The "two 32-bit key values" signature is used for both of those lock functions.
The first key is an arbitrary number I've assigned to each entity type we
store in memcached. Those numbers are stored in entities.json. I've moved the
cache prefixes there too, for consistency. The second key is just the id of the
entity being added to or deleted from the cache.

Keeping the locking in a central place meant making it so that `EntityCache`
and `CoreEntityCache` shared the same `_add_to_cache` method. They were unified
by factoring out a `_create_cache_entries` method that could be extended as
necessary.

Since an `_add_to_cache` method was the only important thing defined by
`EntityCache`, I was able to remove that package entirely, and move
`EntityCacheBase` in its place.