MBS-8801: Fix spurious group-with-gender error
An artist edit can remove the artist's type or gender by setting it to
NULL (represented by undef in Perl). On the other hand, it can leave
fields unchanged (represented in Perl by the absence of a hash entry).
Because of a test for definedness instead of existence of a key before
falling back to the current state of the artist, it could happen that
an edit was rejected for resulting in a group artist with a gender
when that was not actually the case.