MBS-8837: rare ISE editing event time
Broken in 6a8e2e6127cbdec3e23590368e78c2490c200768

This ISE can occur if the time being set is already the current value.

The source of the bug is due to the change in parameter order associated with
Text::Diff3's merge subroutine.

Before the referenced commit, CheckForConflicts hashed the results of
merge_time in the order (ancestor, current, new), which would be something like
[['19:30', '19:30'], ['19:00', (DateTime object)], ['19:00', '19:00]], where
the first value in each array ref is the key. Since the last two array refs
have the same key, the second would override the first.

After 6a8e2e6, those array refs are hashed in the order new, ancestor, current,
which means the DateTime object is the result.

The fix here is to revert the change in the order they're hashed (new should be
last), and always return a formatted string for the current value, not a
DateTime object, since the object isn't stringified to the same format that
ancestor and new are in.