Store server.js request params in redis
This is concerning the data we send to root/server.js in order to render
whatever page we're rendering.

Sending a body along with GET requests isn't really up to spec, so I think we
ought to be doing something that doesn't rely on the nonconforming behavior of
whatever modules we happen to be currently using.

This stores the JSON needed by the page in redis temporarily (keyed by URL +
a random token).

This probably negligibly increases the page load time.