Truncate long stack traces in catalyst log
Since these are currently unbounded, a timeout that occurs on a method with
an enourmous parameter list may generate megabytes of logs from a single
request.

The stack traces tend to become more and more useless the further they go
down, anyway, but if this needs to be configurable then we can make it so.