MBS-8716/MBS-8717: Move edit data as JSONB to separate table
The edit data constitutes a large part of the edit table, making
operations on it slow and requiring more memory for caching. Move it
into a separate edit_data table that can be joined only where
necessary. Also use the more efficient JSONB type for storing it.

With this, the PL/Perl language and the JSON extraction function
written in PL/Perl become unnecessary.