MBS-8940: Make banner dismissal temporary for banner editors
Since banner message editors tend to forget about messages that they
have dismissed previously and do not remove them, limit the effect of
dismissing a banner to 20 hours for them. After that time, the banner
will be reshown, but can be dismissed again (for another 20 hours).
For regular users, the dismissal remains permanent as before, i.e. the
banner will be hidden until a new banner message is set.