MBS-8943: Let direct search work even in read-only mode
Direct search used to fail when the server was in read-only mode due
to its use of run_in_transaction. Change the implementation of
read-only mode in the database wrapper: No longer abort every attempt
to send a COMMIT, even when the transaction didn't change any data;
instead, set all transactions to read-only mode immediately, causing
PostgreSQL to raise an error if and only if a write command is to be
executed within the transaction.