MBS-8806: Don't cache non-existent ACs as empty
When Data::ArtistCredit::get_by_ids was called with IDs that didn't
actually exist, it returned empty ArtistCredit entities for them.
Those were then automatically put into the cache and persisted there.
When an artist credit with the relevant ID later actually was created,
the empty artist credit was displayed in its place until the old cache
entry expired (MBS-8806). It is unclear how get_by_ids could be called
with a not-yet-existing ID.

Only create and return an ArtistCredit entity for IDs actually found
in the database. Also emit a stack trace when a non-existent ID is
encountered in order to aid further debugging.