Correct a translated string
Unfortunately, the string extractor for templates is not sophisticated
enough to detect strings that are split into multiple parts and
combined with the concatenation operator (_). It only finds the first
part, which is then used as the ID in the translation files, breaking
gettext.

Undo a split made in commit c4c1d1ae74b05528de6b73a1294524dc5402bd01.