Use machine-readable Postgres version
When comparing Postgres versions, it is recommended to use the
server_version_num pseudo-setting, which provides a numerical
representation of the version, instead of the version() function,
whose result is intended for human consumption. See also
https://www.postgresql.org/message-id/4422.1465937807%40sss.pgh.pa.us

Accordingly, use the numerical version; this allows us to remove some
parsing code, too.