MBS-9031: Replace memcached with redis
Redis can do everything we're using memcached for (and more), so this
simplifies our setup by dropping memcached as a dependency.

I didn't see any point in using Cache::Redis, because we'd have to write a
wrapper around it anyway (it handles serialization on its own, doesn't have
delete/*_multi methods, etc.), and Cache::Redis itself is already a very small
wrapper around the Redis module. If we were to use Cache::Redis, we'd have to
add a flag to MusicBrainz::Server::CacheWrapper to disable serialization, and
add expiration time support to set_multi. However, another advantage of having
a new wrapper is that we can use Redis-specific features like mget in
get_multi and command pipelining in set_multi.