Support static resource domains (staticbrainz)
The majority of the changes are to do with images. Images are now copied
to root/static/build along with JS and CSS, their filenames appended with a
hash, and all references to them looked up in rev-manifest.json.

This makes it trivial to copy all assets to another server and run zopfli on
them, for example.