Remove/merge redundant URL cleanup tests
Prior to this pull request, clean URL was not checked to match the
same relationship type as the input URL.  This is now auto-checked.

This patch removes and merges some tests which are redundant by
their entity type, relationship type, and input/clean URL patterns,
all at once.

(This actually removes 10 tests.)