MBS-5733: Allow to validate otherdatabases URLs
Previously, limited validation rules applied to `otherdatabases` URLs.

With this patch, validation rules for any specific other database can
be added through a `validate` property in a dedicated CLEANUPS entry.

This patch moves the validation of (the only validated other database)
Soundtrack Collector links to a specific CLEANUPS entry for this
website.  It removes all hard-coded `otherdatabases` validation rules
to rely on the validation process introduced in the previous commit.

Note: It does neither fix nor complete validation rules for Soundtrack
Collector URLs.  This will be fixed as an example in the next commit.