Strengthen Soundtrack Collector URLs validation
Prior to this patch, SC soundtrack URLs were blocked for artist
entities, SC composer URLs were blocked for release_group entities,
and both soundtrack and composer were blocked for release entities.

With this patch, the above cases are still blocked and additionally:
- only clean SC composer URLs are allowed for artist entities
- only clean SC soundtrack URLs are allowed for release_group entities
  (all other SC URLs are blocked for these two entity types)
- all SC URLs are blocked for all other entity types
- all SC URLs are blocked for all other link types

This comes as an example for `otherdatabases` validation (MBS-5733).