Remove Catalyst::Plugin::AutoRestart
Remove this module for a few reasons:

 1. Starlet already kills workers for us between a configured min and max
    number of requests.

 2. Proc::ProcessTable can't compile against musl libc, which is used in Alpine
    Linux (though this isn't really an issue anymore, because we switched our
    docker images to use phusion/baseimage).

 3. We can finally close MBS-4071 (Proc::ProcessTable does not scale).

With the settings currently used in production, our logs confirm that workers
always get killed after 60 requests (the minimum they're required to handle),
because they always exceed a memory size of 350000000 at that point. We can
change the Starlet settings min- and max-reqs-per-child to approximate that
behavior.