Added output_success() to the JSON serializer.
The web service processing uses the Accept: header (or, for some types of
access the 'fmt' parameter) to set up a serializer for its response.
In the case of a successful submission (e.g. submitting barcodes or ISRCs),
it calls an 'output_success' method to report that the submission succeeded.
However, the JSON serializer did not have such a method, causing a server
error when such a submission had the Accept: header set to JSON.