Make Dockerfiles at build time
Instead of storing them in the repository, and having to regenerate them every
commit. The only reason that was necessary was because Docker Hub's automated
builds feature couldn't execute any sort of script before building images; the
Dockerfiles had to already exist in the repo. That feature isn't really useful
to us anymore.