Generate GIT_INFO at build time
Embed GIT_INFO directly into the Dockerfile, because the images themselves
don't contain `git` or any .git folder.

Remove the newline in the MSG part, because newlines have to be escaped in the
Dockerfile, and a colon works just as well.

Change the command used to get the branch name to be slightly more portable.
Using sed for the '*' replacement in bash would be a hassle due to differences
in BSD and GNU sed.