Deploy static files directly to build/, not /tmp/
Ideally, /tmp/staticbrainz would have been a Docker volume prior to this patch,
so that the build state (and file mtimes) were preserved even after the
container was removed. But root/static/build/ is already a volume (because
the rev-manifest.json needs to be shared with the template-renderer). And
there's no reason to have two volumes storing the same files; we can just have
the .gz files go in root/static/build/ as well.