MBS-9136: Fix the TracksWithSequenceIssues report
The TracksWithSequenceIssues report had a number of syntax errors in
its SQL and therefore failed to run. Even if it had run, the third
part of the test would have included all mediums with pre-gap tracks
because it did not account for the additional track.

Rewrite the SQL for the WHERE clause to fix these problems. At the
same time, remove the additional check that pre-gap tracks can only be
used with certain medium types; while having pre-gap tracks on, e.g.,
vinyl is also wrong, this is a very different type of problem and
should have its own report.

(Because the code from commit e002c4d4830502b314e42d178eba8f29ef8e1cd2
never actually worked, this can't be considered a regression.)