MBS-9138: Avoid redundant requests for image-less entities
For entities supporting images, when there is no image to display
(based on any Wikimedia Commons or Wikidata URLs), JavaScript code was
still generated for the client to retrieve an image from the
.../commons-image endpoint. That endpoint then had to load and check
the same URLs, only to find again that no image exists; it even did an
unnecessary call to the Wikimedia API for an image with an empty
title. This may have caused considerable additional and useless load.

When no image is to be found, return early without querying Wikimedia
and without prompting the client to access the commons-image endpoint.