Don't suppress cron output
The (OUTPUT=`foo` || echo "$OUTPUT") lines would only echo the stdout of `foo`
if it returned a non-zero exit code. This seems to have been so that cron
wouldn't send unnecessary emails if everything was successful -- assuming
everything was silent otherwise. In practice, this was broken because we didn't
use the "OUTPUT=" trick or 2>&1 consistently.

In our NewHost setup, we redirect the stdout/stderr of all these scripts to
log files (with the intention of storing the entire output), so it's wrong to
silence the output like this. There's no easy way to see the progress of a
currently-running script otherwise. If cron is sending unnecessary emails, it
should be silenced in the crontab, not here.