Run hourly sitemaps script at :30
This is how it was in the old setup. Running it at :00 doesn't work correctly
because the last replication packet hasn't been pushed yet. So it currently
processes two replication packets every other hour, instead of one per hour.
(Two because processing the first packet gives enough time for the second to be
generated.)