Fix bug related to edit type extra documentation
Bitmap noticed a method call inside a (single-quote) string, which of
course can't work. It seems nobody ever attempted to add extra
documentation about specific edit types to the wiki, so this issue
was never encountered; still, fix it so that it may work in case it
will be used in the future.