MBS-9175: Include non-TOC mediums in "Attach TOC" search
When searching for a medium to attach a TOC to, only mediums of a type
that supports TOCs (such as CD) are candidates. However, it is
desirable to show even other mediums with the right number of tracks:
Not doing so confuses users ("I'm dead sure this artist has a release
with 5 tracks!"), and mediums of the wrong type may serve as model for
a release duplicate in the right type. For these reasons, a search by
release title has always returned all mediums with matching track
count (selecting one of a non-TOC type was disabled in the UI); a
search by artist, however, did not include mediums of types that don't
support TOCs.

Make this consistent by returning non-TOC mediums when searching by
artist, too. They will be disabled UI-wise in the same way as they
already were for a search by release title.