MBS-9186: Gulpfile fails to build scripts in website containers
The Vinyl object mentioned in the comment in the gulpfile wasn't
actually being used anywhere. So now it is. And instead of having it
being required based on the build path (which can be variable), it now
uses an ID, "rev-manifest.json". (Note, if a string passed to `require`
doesn't look like a path, yarb will treat it as a module name relative
to node_modules/, or as a global ID assigned via a call to `expose`.)