Factor out role to traverse FKs
Code from Sitemap::Incremental that would be useful for MBS-8736 is
factored out into a role. The use is that we'll want to generate a dump
that contains all linked data for the entities we select, in addition to
satisfying FK constraints. The latter isn't essential, but still a very
nice property to have, because it makes the data easier to use.

The memoize() calls on `get_foreign_keys` and `get_primary_keys` were
replaced with an explicit cache, because memoize() doesn't seem to work
with roles; consuming packages can't find the memoized methods.